+++ deprecated-react-native-prop-types/DeprecatedViewPropTypes.js
@@ -16,23 +16,42 @@ const {
   AccessibilityRolePropType,
   AccessibilityStatePropType,
   AccessibilityValuePropType,
+  RolePropType,
 } = require('./DeprecatedViewAccessibility');
 const DeprecatedViewStylePropTypes = require('./DeprecatedViewStylePropTypes');
 const PropTypes = require('prop-types');
 
-const stylePropType = DeprecatedStyleSheetPropType(
-  DeprecatedViewStylePropTypes,
-);
-
 /**
  * @see facebook/react-native/Libraries/Components/View/ViewPropTypes.js
  */
-module.exports = {
+const DeprecatedViewPropTypes = {
+  'aria-busy': PropTypes.bool,
+  'aria-checked': PropTypes.oneOfType([
+    PropTypes.bool,
+    PropTypes.oneOf(['mixed']),
+  ]),
+  'aria-disabled': PropTypes.bool,
+  'aria-expanded': PropTypes.bool,
+  'aria-hidden': PropTypes.bool,
+  'aria-label': PropTypes.string,
+  'aria-labelledby': PropTypes.string,
+  'aria-live': 'polite' | 'assertive' | 'off',
+  'aria-modal': PropTypes.bool,
+  'aria-selected': PropTypes.bool,
+  'aria-valuemax': PropTypes.number,
+  'aria-valuemin': PropTypes.number,
+  'aria-valuenow': PropTypes.number,
+  'aria-valuetext': PropTypes.string,
   accessibilityActions: PropTypes.arrayOf(AccessibilityActionInfoPropType),
   accessibilityElementsHidden: PropTypes.bool,
   accessibilityHint: PropTypes.string,
   accessibilityIgnoresInvertColors: PropTypes.bool,
   accessibilityLabel: PropTypes.node,
+  accessibilityLabelledBy: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.arrayOf(PropTypes.string),
+  ]),
+  accessibilityLanguage: PropTypes.string,
   accessibilityLiveRegion: PropTypes.oneOf(['assertive', 'none', 'polite']),
   accessibilityRole: AccessibilityRolePropType,
   accessibilityState: AccessibilityStatePropType,
@@ -40,7 +59,10 @@ module.exports = {
   accessibilityViewIsModal: PropTypes.bool,
   accessible: PropTypes.bool,
   collapsable: PropTypes.bool,
-  hitSlop: DeprecatedEdgeInsetsPropType,
+  hitSlop: PropTypes.oneOfType([
+    DeprecatedEdgeInsetsPropType,
+    PropTypes.number,
+  ]),
   importantForAccessibility: PropTypes.oneOf([
     'auto',
     'no',
@@ -66,7 +88,11 @@ module.exports = {
   pointerEvents: PropTypes.oneOf(['auto', 'box-none', 'box-only', 'none']),
   removeClippedSubviews: PropTypes.bool,
   renderToHardwareTextureAndroid: PropTypes.bool,
+  role: RolePropType,
   shouldRasterizeIOS: PropTypes.bool,
-  style: stylePropType,
+  style: DeprecatedStyleSheetPropType(DeprecatedViewStylePropTypes),
+  tabIndex: PropTypes.oneOf([0, -1]),
   testID: PropTypes.string,
 };
+
+module.exports = DeprecatedViewPropTypes;